update the repo on 0.19 changes
Internal work trying to make the API work for the broad range of cases
that people need was not successful. My feeling is that effect managers
may not be a great fit for web sockets because they do not have great
mechanisms for uniquely identifying resources. Perhaps this can be
solved in the compiler by changing the effect manager design, but I
think it'd be a really big project to solve the root identity question,
and I think it'd better to prioritize other projects first. Using ports
is not ideal, but it works, so knowing that the project could take a
few months, there are other things that seem like more promising time
investments for now.